Message model channged to struct. MessageFactory implemented

+++ src/Mavlink/Packets/PacketValidator.cs
@@ -42,10 +42,10 @@ namespace Mavlink.Packets
             crc = X25CrcAccumulate(packet.SequenceNumber, crc);
             crc = X25CrcAccumulate(packet.SystemId, crc);
             crc = X25CrcAccumulate(packet.ComponentId, crc);
-            crc = X25CrcAccumulate(packet.MessageId, crc);
+            crc = X25CrcAccumulate((byte)packet.MessageId, crc);
             crc = packet.Payload.Aggregate(crc, (current, payloadByte) => X25CrcAccumulate(payloadByte, current));
 
-            return X25CrcAccumulate(GetCrcForMessageId(packet.MessageId), crc);
+            return X25CrcAccumulate(GetCrcForMessageId((byte)packet.MessageId), crc);
         }
 
         private static int X25CrcAccumulate(byte packetByte, int crc)
